requirements.txt
# Core Flask dependencies
Flask==3.0.0
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.0

# Cheminformatics
rdkit==2023.9.1

# Machine Learning & AI
transformers==4.41.1
torch==2.2.2
sentence-transformers==2.7.0
peft==0.10.0
google-generativeai==0.8.5
scikit-learn==1.3.2

# Graph visualization
networkx==3.2.1
pyvis==0.3.2

# FAISS for vector search
faiss-cpu==1.8.0

# LLM APIs
groq==0.4.1

# Data processing
feedparser==6.0.10

# Production WSGI Server
gunicorn==21.2.0

# OCR & Computer Vision (for Prescription Processing)
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0
Pillow==10.1.0
scikit-image==0.22.0

# Medical NER & NLP
spacy==3.7.2
# Note: Install scispacy models separately with:
# pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz

# Text Processing & Fuzzy Matching (for OCR error correction)
python-Levenshtein==0.23.0
rapidfuzz==3.5.2

# File uploads
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
